Display dates on posts
This renders in two steps: the server writes the human-readable date
in UTC (because it neither knows nor wants to know about the client's
time zone), but also writes a time value (millisecond timestamp) for
Javascript to read. On page load, Javascript reads these and rerenders
the dates in the local timezone. (The same system will be used for
chat message dates, once chat messages exist on the server.)