fix bug that appears in recent Flask
Some recent version of Flask changed the default behavior of the .json
property to raise a 400 error for a request that wasn't actually JSON,
rather than simply returning None. This caused a failure when running
a CSRF-checked POST view that accepted form-urlencoded rather than
JSON. Update to allow both again.