Code support for dark mode
Instead of refreshing on dark mode switch, the HTMX endpoint now
triggers a set-dark-mode event. The listener for this sets the
appropriate data-theme attribute on the root element, then triggers
reinit of the rich-text editors.

Rich text editor objects are now used through the RTEWrapper object.
This 1. centralizes the code for RTE initialization, and 2. keeps a
master list of RTE instances currently active. RTEs can be seamlessly
reinitialized while saving their contents; this is used to switch
between light and dark mode.

The mode switch element is now a small alpine component that renders
itself as appropriate depending on the current setting.