Update code for sanitizing HTML
Every data structure that describes a post now uses the PostHTMLText
class; whenever HTML is read in (whether from the client or from the
DB) it's sanitized and verified.

Now has the from_user_input() and output_html() methods for transforms
other than plain sanitization. This is useful primarily for allowing
color; the only way to set that is a style= attribute, which isn't
safe, so we 1. accept style= attributes from the user, 2. verify the
color setting and move it to a (safe) data-color attribute or similar,
then 3. move the data-color attribute back to style on output. The
invariant is that stored HTML is always 100% safe; the output_html
method may produce output with unsafe tags or attrs, but is
responsible for ensuring they're used safely.